chore(tests): move load_event helper to global utils
