+++ tests/functional/data_classes/test_amazon_mq.py
@@ -2,7 +2,7 @@ from typing import Dict
 
 from aws_lambda_powertools.utilities.data_classes.active_mq_event import ActiveMQEvent, ActiveMQMessage
 from aws_lambda_powertools.utilities.data_classes.rabbit_mq_event import BasicProperties, RabbitMessage, RabbitMQEvent
-from tests.functional.utils import load_event
+from tests.utils import load_event
 
 
 def test_active_mq_event():
