+++ tests/functional/data_classes/test_lambda_function_url.py
@@ -1,5 +1,5 @@
 from aws_lambda_powertools.utilities.data_classes import LambdaFunctionUrlEvent
-from tests.functional.utils import load_event
+from tests.utils import load_event
 
 
 def test_lambda_function_url_event():
