+++ tests/functional/event_handler/test_lambda_function_url.py
@@ -1,6 +1,6 @@
 from aws_lambda_powertools.event_handler import LambdaFunctionUrlResolver, Response, content_types
 from aws_lambda_powertools.utilities.data_classes import LambdaFunctionUrlEvent
-from tests.functional.utils import load_event
+from tests.utils import load_event
 
 
 def test_lambda_function_url_event():
