+++ tests/functional/parser/test_alb.py
@@ -3,7 +3,7 @@ import pytest
 from aws_lambda_powertools.utilities.parser import ValidationError, event_parser
 from aws_lambda_powertools.utilities.parser.models import AlbModel
 from aws_lambda_powertools.utilities.typing import LambdaContext
-from tests.functional.utils import load_event
+from tests.utils import load_event
 
 
 @event_parser(model=AlbModel)
