+++ tests/functional/parser/test_apigw.py
@@ -5,7 +5,7 @@ from aws_lambda_powertools.utilities.parser import envelopes, event_parser, pars
 from aws_lambda_powertools.utilities.parser.models import APIGatewayProxyEventModel
 from aws_lambda_powertools.utilities.typing import LambdaContext
 from tests.functional.parser.schemas import MyApiGatewayBusiness
-from tests.functional.utils import load_event
+from tests.utils import load_event
 
 
 @event_parser(model=MyApiGatewayBusiness, envelope=envelopes.ApiGatewayEnvelope)
