+++ tests/functional/parser/test_cloudwatch.py
@@ -9,7 +9,7 @@ from aws_lambda_powertools.utilities.parser import ValidationError, envelopes, e
 from aws_lambda_powertools.utilities.parser.models import CloudWatchLogsLogEvent, CloudWatchLogsModel
 from aws_lambda_powertools.utilities.typing import LambdaContext
 from tests.functional.parser.schemas import MyCloudWatchBusiness
-from tests.functional.utils import load_event
+from tests.utils import load_event
 
 
 @event_parser(model=MyCloudWatchBusiness, envelope=envelopes.CloudWatchLogsEnvelope)
