+++ tests/functional/parser/test_sqs.py
@@ -5,8 +5,8 @@ import pytest
 from aws_lambda_powertools.utilities.parser import ValidationError, envelopes, event_parser
 from aws_lambda_powertools.utilities.typing import LambdaContext
 from tests.functional.parser.schemas import MyAdvancedSqsBusiness, MySqsBusiness
-from tests.functional.utils import load_event
 from tests.functional.validator.conftest import sqs_event  # noqa: F401
+from tests.utils import load_event
 
 
 @event_parser(model=MySqsBusiness, envelope=envelopes.SqsEnvelope)
